changes to overlap detection to reduce memory footprint
now instead of transferring all remote subdivision data up front, a fixed number are transferred
at a time. as subdivisions come in, they are operated on, then deallocated, and new transfers are
started as needed. thus memory usage is limited to the number of active transfers. this may also
have a side benefit of hiding the cost of communication, as it is now overlapped with computation